I want to upload some data to my mongo db in azure. The data will contain the following fields: Name: "Rex" Score: 1234 Time: 64 Mode: "Hard"
We will amend the game so that when the player wins or loses, during the end game dialog they will have the ability to add their name and submit the score manually only. Their name should be remembered in local storage so the next time the dialog appears, the name is autopopulated and can be changed if needed. The name should be sanitised for safety before being added to the database. Only alphanumerics allowed.

How do I do this? I will need step by step instructions. The uri for my mongo server is: xxx

if a username and password is required, we need a secure way to obtain this using secrets in azure or something secure. We do not want to add these details in code.